Implementation of ALF-6829 Popularity of a node (RatingService).
This check-in is actually a fairly generic support for rolled up property values within rating schemes.
Popularity is currently the only concrete rollup in the system, but it should be possible to add more without too
much difficulty.

Given the published API for the rating service, we have added rollups for "ratingCount" and "ratingTotal" for the built-in rating schemes (likes). Therefore searching/sorting on either cm:likesRatingSchemeCount or cm:likesRatingSchemeTotal, both of which are defined in an aspect cm:likesRatingSchemeRollups should give popularity for the likes rating scheme.


Additions to the content model.
  I have added aspects, properties for the built-in rollups, namely count and total.

Spring changes
  The rating scheme(s) now have rollups injected into them. These rollups are algorithm classes that know how to calculate the rolled up props.
  Naming conventions are used to determine the output property and are captured in RatingRollupNamingConventionsUtil.java.

API changes
  New methods: RatingService.getRatingRollup(NodeRef, String, String)
  & RatingScheme.getPropertyRollups()

Changes to the RatingService implementation and JUnit test code.

To roll your own rating rollups:
  To add your own rating rollup, you need to reuse one of the existing AbstractRatingRollupAlgorithm subclasses or create a new one.
  That class will produce the value of the rolled up property by iterating the ratings in the given rating scheme for a given node.
  You need to inject your rollup algorithm into the rating scheme bean in rating-services-context.xml
  On applyRating() and removeRating() the RatingService will check the rating scheme for any registered rollup algorithms and will have them recalculated.
  It will then use the naming conventions described in the JIRA and in RatingRollupNamingConventionsUtil to get aspect/property names for each of the rollups.
  It will then add the property value in the normal way.
  Therefore, you need to extend the content model to include the expected aspect and property values. Follow likesRatingScheme for a sample.