ALF-684: Ftp requests to the IPv6 resolved hostname are failing
The ftp.ipv6.enabled property was removed, and a ServerSocket is opened without InetAddress parameter. It allows Java to determine if a IPv6 is used and bind "::" any local address to the server socket. It allows to connect to the Alfresco FTP using both IPv4 and IPv6 addresses regardless to server OS.