Fix for ALF-5084, ALF-5644 and general thumbnail improvements in the area of failing thumbnails.
This check in adds extra functionality to the create-thumbnail action in order to improve how Alfresco handles failing thumbnails.
The requirements agreed between Services & Prod Mgt were
 - When thumbnails fail we should limit how often they are retried. So the system should remember when a thumbnail has failed for a node. 
 - The retries will be initiated in the same way they are now: by users navigating to a page that asks for the thumbnail (lazy creation), primarily doclib. 
 - Retries will occur at a throttled rate (all the below parameters to be configurable by sys admin, but not yet by JMX). 
 - Initially, retries will be limited to a maximum rate e.g. 1 per minute. 
 - These retries will be allowed up to a maximum count e.g. 2 failed attempts. 
 - After that count is reached, the content will be considered "difficult" and thumbnail creation should be retried at a much reduced maximum frequency e.g. once a week 
 - Retries of difficult content shall be enabled by default, but can be disabled.

This check-in adds a new aspect and a type to the content model. These hold the thumbnail failure information.
The create-thumbnail action now has a compensating action which is run when it fails.
  AddFailedThumbnailActionExecuter decorates the content node with failure information.
create-thumbnail now has an action-condition, which means that it will only run if a node is "eligible" for a thumbnail creation attempt.
Eligibility is controlled by some new properties and is configurable. Default behaviour is as described above.

There should be no change in behaviour for most content nodes in Alfresco.
However, for those documents which fail to produce doclib or webpreview thumbnails - and most especially those which consume 100% cpu for some time before failing - we expect to see the same behaviour initially on upload (cpu grind as Alfresco or a 3rd party library attempts to produce a thumbnail), but that as time passes, Alfresco should prevent thumbnail creation that will likely not succeed. This should save cpu on the repo server.

And tests. Of course.