Services-layer fix for ALF-8036. Incorrect permissions copied when copying folder with permissions from one site to another.
UI will have to make changes in the UI layer in order to use new Services methods, thus completing the fix.

This check-in adds support in the SiteService for moving and copying site-contained nodes. New move/copy methods are available in the Java Foundation API and in the JavaScript API which mimic the signatures of those in the NodeService and CopyService respectively. These methods simply delegate to the back-end services and so behaviours should be the same as before if they are used in place of the node and copy service methods. With one exception:
The SiteService wrapper methods detect when the relocated node has been copied/moved between two different Share sites and clears permissions from the relocated node (and its primary descendants) that refer to the previous site.

This has not been implemented using policies/behaviours as we can't bind the behaviour to any particular content class. Nodes of any type could be copied/moved between sites.