More on ALF-7004. The notification templates (used in the send email action of a rule) now follow the look and feel of the mockups Linton did originally.
For this to work the MailActionExcecuter action needed the shareUrl, as all other email sending processes go through this action they no longer need to pass the shareUrl in. There was some inconsistency in how the URL was structured (sometimes with a trailing slash, sometimes not) so this has been cleaned up, which did mean more template updates.

As before, this will require a clean database to see the updated templates.