ALF-8615: ACTVT 106: Alfresco WAR includes all the necessary assets to run the Activiti Admin UI ALF-8616: ACTVT 107: Enterprise build overlays required Activiti Admin UI files
NOTE: To get the Activiti Admin UI you need to do an Enterprise build. If you're using the exploded targets there are no extra steps. 

If you are using the default, WAR based targets you will first need to run "ant include-activiti-admin-ui" in the 'enterpriseprojects' folder. Due to the way we build our WAR file the original source file has to be changed, there is a potential this may be accidentally comitted, hence the extra manual step. In a continuous build this step is automated and the changes reverted when a new build is performed just as all the other enterprise overlay files are.

To see the Activiti Admin UI go to http://localhost:8080/alfresco/activiti-admin. If you login into Explorer first you'll go to straight to the app, otherwise a login dialog will be displayed.