Implementation of ALF-8969 Lucene removal: Blog webscripts.   ** Checking this in on Swift branch, as I have the work there. Will merge to HEAD.
The blog webscript controllers have been ported from JavaScript to Java.
A new foundation service, the BlogService has been added and the impls of the webscript controllers delegate into that service, thus encapsulating business logic within the service.
The API for this service is based on the requirements of the existing webscripts, but is for the most part a 'sensible' API. One controller (blogposts.get.js) had very domain-specific requirements (get all of my drafts and all published posts) and it is implemented as a deprecated public method on the service.
The API is not complete, but represents a good starting point for any future feature development.
The various Lucene queries have been replaced with calls to the nodeservice (as an impl detail within the BlogService) which get all blog post nodes and then post-filter them based on property values, aspect/property presence etc. This will be refactored into a CannedQuery in a subsequent check-in.
 
I've written new test cases aimed at this API & have extended the REST API tests.