THOR-25 Copy and refactor Account Service from SambaJAM I've copied the AccountService and related classes and spring config from SambaJam into Alfresco. I've deleted most of the code which is not applicable to Alfresco.
I've also added some basic test cases - a test to start up the module & a test that uses a dummy AccountService to create/delete accounts. These are the seeds of more realistic tests that will come later.
I'm following my standard pattern of a TestSuite called AllCloudTests, which won't be run by the build but can be used in Eclipse to run all tests relevant to this feature.

Also various refactorings, polishings, adjustments throughout the code. More to come, obviously.