Completed ALF-9392: More workflow metrics are required on the Workflow Console Admin Tool
The Workflow console tool has switched to a "generic form tool" and now reports Workflow metrics from the "Alfresco:Name=WorkflowInformation" MBean, the engines enabled state, number of definitions, number of workflows and number of tasks are displayed.

The link to the "Activiti Workflow Console" is then placed under the form.