THOR-49. Implementation of Reset Forgotten Password workflow. Added 3 new email templates to the bootstrap for Thor. New bpmn2.0 xml for the workflow definition. (No diagram yet). Various Activiti Delegate classes 2 new REST webscripts: request-reset-password.post.desc.xml to init the workflow & change-password to update the password. And I've hidden the reset-password workflow from Share. Added *.rcq files for the 2 REST calls. And REST-based tests for all flows.
Still to do: add a timeout to the workflow and an email to notify of changed password (but I've no reqmnts for these).
Also. I need to check that the password is actually updated! I'm not convinced my test covers it properly and my testing using a real repo is slow, but otherwise complete.