THOR-659: Caching Content Store blocked readers
Corrects method from Thread.run() to Thread.start() so that the code is genuinely run in a fresh thread. Also changed the thread type to daemon so that it doesn't hamper JVM shutdowns.

A further change for this issue will be to switch to a single reused thread rather than a new thread per cleanup request.