Partial fix for MOBILE-30: Propose saving to "Downloads" when a document sent via QO "Save Back" can't be saved
The fix does as the bug suggests but there is an intermittent issue still where the failed response goes through a different code path and therefore does not offer the prompt to save to Downloads. I also need to have the wording of the messages reviewed before the bug can be fully resolved.