Cloud SPP: remove remote-api dependency from repository.
A dependency on the remote-api project had accidentally crept in to the repository. This was due to explicitly loading the remote-api-context.xml file in the application-context.xml of the repository project. This meant that when running integration tests within the IDE for example, the remote-api-context.xml file could not be found as it is in a different project that the repository should not depend upon.

Loading the remote-api-context.xml file has moved to the Thor project where it belongs (note, that in HEAD this is where it has always been, web-application-context.xml)