ALF-15889: decouple TransactionalCache from EhCache
It isn't 100% clear whether widening the scope to Throwable is ok at this point. However, the cache related methods may produce exceptions other than net.sf.ehcache.CacheException, for example EhCacheAdapter.clear() could produce an IllegalStateException and this was not caught by the previous catch(CacheException) block. An alternative approach was to create our own CacheException however this would potentially have to be thrown by every SimpleCache cache method for every implementation and given the IllegalStateException reference above, it is not obvious whether just ehcache CacheExceptions should be translated to alfresco CacheExceptions or whether all exceptions would be wrapped/translated in this way. For the moment, this modification will do - may be reviewed later.