Cloud SPP: really fix import order.
I was trying to avoid doing this and had thought that I had fixed the ordering requirements of Thor being loaded AFTER Sharepoint. However the previous fix of explicitly loading the sharepoint context at the top of the thor context did not work because spring imports can happen more than once. The only reliable fix is to change the name of the sharpoint module-context.xml file so that it is not loaded automatically by 'classpath*:' urls.

This means that merging into/from cloud1_sp will be harder than it was.