Additional improvements for ALF-17586. Support for dynamically finding data keys.
Added extra commenting to the CompositeAttribute tests to help MikeF make sense of the API.

Also added in extra test code that retrieves the data 'keys' from a composite attribute value.
It turns out that this was not previously possible - which may explain Mike's trouble with the API.

Provided an additional method ScriptMBeanAttributeCompositeValue.getDataKeys() which can be used to access the 'keys' in a composite attribute value.
For reasons best known to Rhino, the inherited keySet() method in HashMap is not available. WTF?