Final major refactor of ScriptableMBeanOperations. For ALF-18396. Added support for the invocation of all MBean operations, including overloaded JMX operations. Given Java and JavaScript's differing views on overloading, this requires a bit of jiggery-pokery.
JMX operations are now exposed in the JavaScript API in two forms. All operations are exposed in a 'type-aware' form. (The trailing underscore is intentional here.)
  bean.operations.doSomething_(['java.lang.String', 'int'], 'hello', 42);
where the JavaScript author must provide an array of Strings representing the Java types of the JMX operation parameters. This is to support the invocation of overloaded JMX operations. Autoboxing and unboxing is not supported. You must specify exactly the right Java type.

A second 'simple' form is available only for JMX operations which are not overloaded with different method signatures varying only by Java type.
  bean.operations.doSomething('hello', 42);

For example, an MBean defining operations doSomething(String) and doSomething(String, String) would be callable in the simple or type-aware form. An MBean defining operations doSomething(String) and doSomething(int) would not be callable in the simple form, only the type-aware form.

And tests for all this.