BM Share: Give currently-active scenario a first shot at generating events
Previously, if the active scenario chose no events, it would not be eligible generating more
events when another scenario was chosen.  Now, the active scenario can generate more events on 
the first pass; if it does not (i.e. it considers itself finished) then it can still be asked
to generate events on an equal footing with all the other scenarios.

It is now possible to have a scenario complete and kick off repeatedly even if it is
the only scenario available.