Fixes lost message receivers.
Previously, when a proxied NullMessenger was being created (with a view to it being upgraded to a HazelcastMessenger later on), the receiver was being discarded and so even after upgrade the HazelcastMessenger was unable to dispatch messages to the receiver. This is now fixed - a NullMessenger looks after the receiver reference, even though it isn't using it, then it is passed on to the HazelcastMessenger when it is created.