Clustering: improved IP address handling.
Repo nodes register themselves in the database using the same IP address as selected by Hazelcast's AddressPicker. Previously it was possible for the node to be registered in the database using a different IP address to the one that Hazelcast was using (when machine has multiple interfaces).

When the addresses were different, JMX reporting would not be reliable: for example, it might say that there was an offline member whereas it was really a cluster member registered with a different IP.

It is now possible to use the interface property:

  alfresco.cluster.interface=10.244.*.*

and this will be resolved to an IP address that is used consistently throughout clustering. Useful for ensuring a different interface is used for clustering than for tomcat request handling for example - and without having to configure each repo node with a specific IP address.