MNT-2676: kerberos: why does alfresco need to connect to the KDC if it has the SPN keytabs and why do we store the SPN passwords
All the code related to SPN passwords for Kerberos authentication and related to the 'kerberos.authentication.cifs.password' and 'kerberos.authentication.http.password' parameters was removed. Receiving of the security keys will be delegated to 'Krb5LoginModule' to be extracted from key tables. All the Kerberos related Web Scripts have been also modified to suite all the changes in configuration