Preparatory work for fixing CLOUD-1754. MailActionExecuterTest failing. It has been very difficult to get reliable test feedback on dealing with the MailActionExecuter. This is largely because of the difference in the way mail actions are handled in cloud and on enterprise. Therefore I have refactored and extended our test code to support fixing the MailActionExecuter.
Part 1: Refactor the JUnit Rules AlfrescoPeople, AlfrescoPerson, AlfrescoTenant so that when a test is run in
Community, Enterprise or Cloud the test 'does the right thing' when creating/tearing down test users.
i.e. On Community/Enterprise, create the user/auth and the person on the single tenant.
In Cloud create the user in system tenant, create the person in their home tenant, precreating the tenant if necessary
and tear the whole thing down again.

I achieved this by having the cm:person-related rules delegate to a spring bean responsible for dealing with test users.
Then I've got a mechanism to let Enterprise override Community and Cloud override Enterprise.
So the Community 'TestUserComponent' just creates/deletes users. Enterprise 'inherits' this behaviour.
The Cloud TestUserComponent deals with the tenant sideshow.
It's all started in the global-integration-test-context.xml.