Fix RepositoryStartStopTest build failure.
Merge error led to this test being included in the files.tests.startup group - removed so that this is the same as for HEAD-QA and clustering_p2. This test requires a full-context to have been run prior to its test run, otherwise it fails with the missing namespace URI dictionary errors as seen in the build failure.