Fix for ALF-18781. Subsystems not restarting after JMX attribute updates via JS API. The fix here was to use AbstractPropertyBackedBean.setProperties() rather than setProperty. Atomically updating attribute values in one call has the side-effect of ensuring attribute value changes across a cluster and starting subsystems.
There is a major limitation there though, in that the setProperties() call only supports JMX attributes whose values are Strings.
So to cut a long story short, the ScriptMBean.saveImpl() method has been refactored to detect whether save() calls are saving string-valued attributes in a subsystem, in which case it will use setProperties(). Otherwise, it uses setProperty() as before.

Tests for same.