ALF-18673: commented code to make expected behaviour clear.
The part of the issue that was truly a bug had previously been fixed (by making Messenger objects ignore messages from themselves by default). The remaining suggested fix (49723@BRANCHES/DEV/BELARUS/HEAD-BUG-FIX-2013_04_11) is in fact not a valid fix and I've replaced the one-line fix with a comment explaining the behaviour.