ALF-19673: WebDav: Odd behavior, when try to create or rename folder
MoveMethod now only creates a zero byte file when source file name not only matches the shuffle regex, but the source file is NOT a folder.

Also added RenameShuffleDetectionTest so that we have at least some testing of the shuffle regex - this will help prevent regressions and provide some future documentation on what cases are handled by the pattern matching.