Merged BRANCHES/DEV/V4.1-BUG-FIX to HEAD:    54638: As part of work investigating ALF-19783 I have created a test case. I'm checking this in on 4.1 bugfix (where the issue is not reported) Merged BRANCHES/DEV/V4.1-BUG-FIX to HEAD:    54640: (RECORD ONLY) (Part of prep work for ALF-19783) Moving a test class to the correct source tree.
Note that in doing the record-only merge above, the file ScriptNodeTest had been created on V4.1-BUG-FIX *and* on HEAD independently.
So I got a tree conflict in merging 54640 to HEAD as this attempted to svn mv ScriptNodeTest on top of HEAD's existing copy of that file.
So instead, I manually resolved the conflict by merging the two files.