Fixes SlowContentStoreTest test breakage. Removes dependency on cachingstore/test-context.xml
This fix essentially denormalizes the various Spring context XML files in the cachingstore directory. A previous fix had aliased the cachingContentStore to fileContentStore, but unfortunately this broke the SlowContentStoreTest tests, as the cachingContentStore created by those tests is not a real/working content store (the backingStore is a SlowContentStore dummy store), so the bootstrap process failed.

By making each of the contexts standalone, the different tests are able to evolve as required without breaking the other tests.