Initial implementation of a queueing mechanism for client-generated events. This checkin adds a major refactoring of the use of jnotify - see JNotifyWatchService for watch management and JNotifyEvent for how the events are now modelled. It adds a 'raw queue' (in EventComponent) which is given all the JNotify events. A RawQueueProcessor then sanitises these events (filtering out noise, recognising create,delete pairs that should be synthetic moves) and adds the sanitised events to a sanitisedEventQueue - again in EventComponent. Another thread (see ClientEventProcessor) then takes from the sanitisedQ, handling the events by calling the Syncer objects as before.
The Syncer objects are unchanged.
The SyncManager now needs to do some component initialisation and wiring and is starting to act like a mini ApplicationContext.

Next commit will integrate all client-side events (GUI button action and timed thread) with the sanitised queue so that every client side
event is handled in a single-threaded, one at a time fashion.