Refactoring of how the client applies changes to jnotify events. This will make it much easier for client devs to develop new cases of 'special behaviour'. I have separated the raw queue processing (meaning the 4 second delay and the basic filtering) from what I'm now calling transformations. (I hate that word. But I hate mapper more. Hmm.) Event transformations get applied to short lists of events - the buffer inside the RawQueueProcessor. These transformations can make no change, or swap out individual events for other synthetic events, or can pair events together or split them. The current examples are create + delete pairs during some moves and secondly the create + modify events we see on Windows during create.
These should now be implemented in a dedicated package, transformations and each transform is one new class.
They run in sequence on the event list, applying every transform the system has configured.
An abstract base class allows for transforms to be easily configured to be OS-specific. If one is only meant for Windows, it'll be ignored on Mac OS X.

I have extracted the existing implementation of TransformCreateDeletePairsToMove into a class of that name. Catchy.
The next example that should be implemented might be TransformCreateModifySetsToCreate, which would be Windows only.

Finally, because the transforms are now in dedicated classes, they can be developed and tested in parallel.

All in all, much better.