Changes to the way the application is shut down - both in production and during automated tests.
The motivation here is to ensure that the application is shut down cleanly. This matters in production of course,
but is also very important during automated test execution where we need to know that there are no threads that remain
running across multiple application launches.

Specifically, I have done the following:
When the SystemTrayIcon is used to shut the application down, it used to simply call System.exit(0); See ACE-400, a separate issue.
It now also calls SyncManager.shutdown() in order to ensure correct behaviour that's consistent with tests.
SyncManager.shutdown() shuts down the two Java executors (thread pools) that are responsible for managing client-side events e.g. jnotify events.
It calls shutdown() as before but then awaits termination to be absolutely sure that those threadpools are gone before returning
from SyncManager.shutdown()
In other words, SyncManager.shutdown() now blocks on complete shutdown of any internal components, which may take a few seconds.