Possible fix for ACE-1957, although not confirmed as I can't reproduce.
We know that the SystemTrayIcon, being an AWT/Swing UI element, must be updated within the Swing thread and *not* within the JavaFX event thread.
Therefore all updates to UI state in the SystemTrayIcon should be wrapped in a SwingUtilities.invokeLater so that they are on the Swing event thread.
I found 2 or 3 cases where that wasn't the case - these were added after Ray Gauss had already fixed all usages of SystemTrayIcon.

On Windows, if you access Swing state from the wrong thread, the JVM seems to muddle through ok. But on Mac, it can result in thread locks and the beach ball of death.

I also removed some dead code and fixed some javadoc.