Add support for Salesforce Console view. Pages now check to see if the view layer is in a console, it then adjusts how links, buttons, and other user initiated actions work.  The AlfDocs component will open in a subtab.  All CRUD type operations occur within that subtab.
This change also required changes to make the alfdocs component usable outside of our standard object set which is a necessary change for allowing custom content modeling. This also required all remoteaction annotated methods (and their classes) to be declared global. These methods are now available as API level methods of the managed package. It should noted in documentation that these methods are not part of the "supported" Public API of the connector.