Initial commit of Java service changes for facet reordering. Part of ACE-1582.
Facet persistence has not been changed, but a property has been added to the folder
which contains the facet nodes. This property 'facetOrder' holds a sequence of strings
which are the ordered facet IDs.
Facets are now returned from the SolrFacetService in a sorted order, where this explicit
'facetOrder' is the primary source for sort order.
The sorting algorithm falls back to using indexes if no order is available.
It then falls back to using alphabetic sorting if no index is available.
These last two scenarios are likely to be corner cases.

Still to do: some enhancements to the Java service, webscripts as REST API endpoints.
             the REST API will provide for reordering of existing facets.