Addition of what I'm calling 'synthetic' facetable properties to SolrFacetService and REST API as part of ACE-2639. A synthetic property is something that is technically *NOT* a property, but which can be treated as such for the purposes of SOLR facetting/filtering. Currently we only support two synthetic properties: size and mimetype, which are defined within any Alfresco property of type d:content. This includes, but is not limited to, the cm:content property in all cm:content types in Alfresco. (Be careful with your cm:content. That name is heavily overloaded.)
This checkin only supports what we need for 5.0. It is expressly not designed for resuse/extension in a general way.
However, I have tried to push the less stable parts of this work into the implementation of the service and REST API
and limit what's in the service interface to the more stable parts of the work.

Also note that I had to add some l10n/i18n key-values for these synthetic properties. I've called the mimetype 'MIME type'
in the l10n'd text as I have translations for that in all languages.