This is a tentative fix for ACE-3671. It's being checked in in order to push one build through bamboo.
The fix was to remove the bootstrap patch for the solrFacetsRootFolder created within Data Dictionary.
We are lucky with this feature in that any time this folder is accessed, it either happens:
1. within a write transaction and so we can lazy-create the folder
2. within a read transaction where the data being returned by the SolrFacetService can be easily simulated.

So we now lazily create the folder when it is first used - reusing the importer bootstrap/view data that the patch had used.
AFAICT, this will create the folder (in the default tenant) during startup and also when any search admin tries to create a new SOLR filter within a non-default tenant.